requirements.txt
aioice==0.10.1
aiortc==1.13.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
av==14.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
cryptography==45.0.3
debugpy==1.8.14
decorator==5.2.1
dnspython==2.7.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.7.1
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.2
idna==3.10
ifaddr==0.2.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==1.41.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
protobuf==6.31.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.1
pylibsrtp==0.12.0
pyOpenSSL==25.1.0
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.1
safetensors==0.5.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
streamlit==1.45.1
streamlit-webrtc==0.62.4
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.3
triton==3.3.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
wcwidth==0.2.13
